fix 10baee1: rm unused test
That test was thought to catch a situation which can't happen.

I thought I'd catch a case when delete was called after pop.
But in order to call delete, one must call mapping first -- but
mapping with a popped key returns pair of nils, which would
make delete crash anyway.
So no need to defensive catching errors of :gb_trees.delete.